further improve verifier part
