+++ camel/utils/__init__.py
@@ -40,6 +40,8 @@ from .commons import (
     to_pascal,
     track_agent,
     with_timeout,
+    TIMEOUT_THRESHOLD,
+    with_timeout,
 )
 from .constants import Constants
 from .deduplication import DeduplicationResult, deduplicate_internally
