+++ camel/utils/commons.py
@@ -39,6 +39,7 @@ from typing import (
     Type,
     TypeVar,
     cast,
+    Awaitable,
 )
 from urllib.parse import urlparse
 
@@ -51,6 +52,9 @@ from camel.types import TaskType
 from .constants import Constants
 
 F = TypeVar('F', bound=Callable[..., Any])
+
+TIMEOUT_THRESHOLD = 180.0  # Default timeout in seconds
+T = TypeVar('T')
 
 logger = logging.getLogger(__name__)
 
@@ -1130,3 +1134,28 @@ def run_async(func: Callable[..., Any]) -> Callable[..., Any]:
         return loop.run_until_complete(func(*args, **kwargs))
 
     return wrapper
+
+
+async def with_timeout(
+    coro: Awaitable[T],
+    timeout: float = TIMEOUT_THRESHOLD,
+    context: str = "operation"
+) -> T:
+    r"""General timeout wrapper for async operations.
+    
+    Args:
+        coro: async operation to be executed
+        timeout: max wait time (seconds)
+        context: context information, used for error messages
+        
+    Returns:
+        result of the async operation
+        
+    Raises:
+        asyncio.TimeoutError: if wait times out
+    """
+    try:
+        return await asyncio.wait_for(coro, timeout=timeout)
+    except asyncio.TimeoutError:
+        logger.error(f"Operation timed out after {timeout}s: {context}")
+        raise asyncio.TimeoutError(f"Timed out while {context}")
